Refactor logging
