+++ Program.cs
@@ -26,8 +26,7 @@ namespace BeanBot
             Support.StartupOperations();
             await LogIntoDiscord();
             await InstantiateCommandServices();
-            _discordClient.Log += LogMessages;
-            _commandService.Log += LogMessages;
+            _discordClient.Log += LogHandler.LogMessages;
             await Task.Delay(-1);
         }
 
@@ -75,34 +74,6 @@ namespace BeanBot
                 LogLevel = LogSeverity.Verbose,
                 MessageCacheSize = 50
             });
-        }
-
-        private Task LogMessages(LogMessage messages)
-        {
-            string formattedMessage = $"Discord:\t{messages.Source.ToString()}\t{messages.Message.ToString()}";
-            switch (messages.Severity)
-            {
-                case LogSeverity.Critical:
-                    Log.Fatal(formattedMessage);
-                    break;
-                case LogSeverity.Error:
-                    Log.Error(formattedMessage);
-                    break;
-                case LogSeverity.Warning:
-                    Log.Warning(formattedMessage);
-                    break;
-                case LogSeverity.Info:
-                    Log.Information(formattedMessage);
-                    break;
-                case LogSeverity.Verbose:
-                    Log.Verbose(formattedMessage);
-                    break;
-                default:
-                    Log.Information($"Log Severity: {messages.Severity}");
-                    Log.Information(formattedMessage);
-                    break;
-            }
-            return Task.CompletedTask;
         }
     }
 }
